requirements.txt
asgiref==3.2.5
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-braces==1.14.0
django-cors-headers==3.2.1
django-extensions==2.2.8
django-heroku==0.3.1
django-oauth-toolkit==1.3.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.9
mysqlclient==1.4.6
oauthlib==3.1.0
psycopg2==2.8.4
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.8
whitenoise==5.0.1
